even more CMake refactoring and moving of DataSave to common dir
