+++ dataformats/multigrid/src/genpixids.cpp
@@ -2,7 +2,7 @@
 
 #define UNUSED __attribute__((unused))
 
-#include <DataSave.h>
+#include <common/DataSave.h>
 #include <MapFile.h>
 #include <cassert>
 #include <getopt.h>
