+++ prototype2/common/FBSerializer.cpp
@@ -3,7 +3,7 @@
 #include <cinttypes>
 #include <common/FBSerializer.h>
 #include <common/Trace.h>
-#include <dataformats/multigrid/inc/DataSave.h>
+#include <common/DataSave.h>
 #include <libs/include/gccintel.h>
 
 #define TIMESIZE 4
